ShopApp/requirements.txt
bleach
django
djangorestframework
djangorestframework-jwt
djangorestframework-simplejwt
mysqlclient
